create a folder for euphoria
